bootutil: Move update-independent code
Move functions from loader.c that are independent from the update type
into a dedicated file. That way it will be easier to provide alternative
loader type without massive code duplication.
This file is intended to be used by loader-like files instead and should
not be included in the bootutil/_priv.h headers.

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>